1 player works, 2 player broken now maybe!
Probably an issue in my main switch statement.

+++ WAndersonLab6/Core/Inc/game_state.h
@@ -9,12 +9,13 @@
 #define INC_GAME_STATE_H_
 
 #include <stdint.h>
-#include "LCD_Driver.h"///////////////// MAYBE NEED TO DELETE
+#include "LCD_Driver.h"
+#include "AI.h"
 
 
-int placeCoin(uint8_t board[6][7], uint8_t col, uint8_t player);
+uint8_t placeCoin(uint8_t board[6][7], uint8_t col, uint8_t player);
 
-int checkWinOrTie(uint8_t board[6][7], uint8_t row, uint8_t col, uint8_t player);
+uint8_t checkWinOrTie(uint8_t board[6][7], uint8_t row, uint8_t col, uint8_t player);
 
 
 #endif /* INC_GAME_STATE_H_ */
